PR Changes: * Move Registration of new Style to StyleSheets.RegisterStyle method
